* RarArchive.m NSArray count returns NSUInteger, not int
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2156 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Zipper/RarArchive.m
@@ -88,7 +88,7 @@ static NSData *_magicBytes;
 
 - (NSArray *)listContents
 {
-	int lineCount, i;
+	NSUInteger lineCount, i;
 	NSString *path = nil;
     
     NSMutableArray *results = [NSMutableArray array];
